+++ samples/elaborate/tetris/model/game.rb
@@ -41,11 +41,12 @@ class Tetris
       SPEED_INITIAL_DELAYS = SPEEDS.each_with_index.inject({}) {|hash, speed_index_pair| hash.merge(speed_index_pair.first => 1.1 - speed_index_pair.last*(0.1)) }
       
       attr_reader :playfield_width, :playfield_height
-      attr_accessor :game_over, :paused, :preview_tetromino, :lines, :score, :level, :high_scores, :beeping, :added_high_score, :show_high_scores, :up_arrow_action, :initial_delay
+      attr_accessor :game_over, :paused, :preview_tetromino, :lines, :score, :level, :high_scores, :beeping, :added_high_score, :show_high_scores, :up_arrow_action, :show_preview_tetromino, :initial_delay
       alias game_over? game_over
       alias paused? paused
       alias beeping? beeping
       alias added_high_score? added_high_score
+      alias show_preview_tetromino? show_preview_tetromino
       
       def initialize(playfield_width = PLAYFIELD_WIDTH, playfield_height = PLAYFIELD_HEIGHT)
         @initial_delay = SPEED_INITIAL_DELAYS[:snail]
@@ -55,6 +56,7 @@ class Tetris
         @show_high_scores = false
         @beeping = true
         @up_arrow_action = :rotate_left
+        @show_preview_tetromino = true
         load_high_scores!
       end